+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- removed validate hook
+- implemented React hook form and updated sign up page. Form logic is now fully in the form components. Added sign in page/form
 - updated routes, added an unfinished sign in page. Set a route outlet for the account pages
 - added form validation and redirect routes based on user authentication
 - Fully implemented Firebase sign in and context auth. Updated react-router-dom imports to correspond with v6 exports. Created private routing for unauthorised access. Created an auth provider wrapper component